backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.1.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
python-decouple==3.8
django-allauth==0.57.0
django-extensions==3.2.3
django-storages==1.14.2
boto3==1.34.0
django-redis==5.4.0
django-ratelimit==4.1.0
django-cleanup==8.0.0
django-imagekit==4.1.0
django-taggit==5.0.0
django-notifications-hq==1.8.2
channels==4.0.0
channels-redis==4.1.0
